Retailored the project scaffold + added first test.
The previous scaffold was designed for writing node modules only,
changed that slightly in order to use it for a web server project. Added
first nodeunit test for creating an S3 bucket if it doesn't exist yet.
Created the website design. Defined the most important routes and
identified the node-modules that will be used later on.