Created multiple scripts that enhance usability.
The overview page can now be dragged up and down in order to scroll
through the gallery of watched animes. The whole website has been
programmed with singlepage capabilities in mind - added a script which
uses that now. Created a quickinfo mechanism for the input form. The
script uses an external API to provide the user with information.